1-fundamentos/1-hello-world-openai.py
from langchain_openai import ChatOpenAI
from dotenv import load_dotenv

load_dotenv()

model = ChatOpenAI(model="gpt-4o-mini", temperature=0.5)
response = model.invoke("Hello, world!")

print(response.content)